pyproject.toml
[project]
name = "llnca"
version = "0.1.0"
description = "large language neural cellular automata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
]
